Initial Commit for my To Do List project
This is a project I wanted to play around with to try and keep my
jQuery skills sharp.  I based it off of a tutorial, located at -
[This link](https://www.youtube.com/watch?v=u4p6x-QH_vI&ab_channel=Mr.WebDesigner)

I'm encountering an error where my FontAwesome icons seem to be
crashing it (this always happens every time I try to use it and I
have yet to figure out why) so that's my next step in debugging.